requirements.txt
streamlit==1.31.1
mistralai==0.1.4
PyPDF2==3.0.1
markdown==3.5.2
pymdown-extensions==10.7
reportlab==4.1.0
markdown2==2.4.12
Pillow==10.2.0
openai==1.12.0
python-dotenv==1.0.1